+++ OgreMain/src/OgreTextureManager.cpp
@@ -344,9 +344,9 @@ namespace Ogre {
 
         // Yellow/black stripes
         const ColourValue black(0, 0, 0), yellow(1, 1, 0);
-        for (size_t y = 0; y < pixels.getHeight(); ++y)
+        for (uint32 y = 0; y < pixels.getHeight(); ++y)
         {
-            for (size_t x = 0; x < pixels.getWidth(); ++x)
+            for (uint32 x = 0; x < pixels.getWidth(); ++x)
             {
                 pixels.setColourAt((((x + y) % 8) < 4) ? black : yellow, x, y, 0);
             }
